langchain-demo/requirements.txt
jupyterlab
python-dotenv
langchain-openai
langchain
openai
langchain-community 
langchain-core
beautifulsoup4
langchain-anthropic
langchain_google_genai 
#langchain_google
huggingface_hub
diffusers
qai-hub
transformers
# qai_hub_models
helper_functions
soundfile
librosa
gradio
timm
inflect
phonemizer
llama_index

pinecone-client
pinecone-datasets
pinecone-text

tiktoken
tqdm

datasets
deepface
pinecone-notebooks

gradio_client